✨ stream -> live로 이름 수정 및 레이아웃 공통화 적용
